docu finetuning and move 3437 specific class oout of legacyinstrument
